Version Final LAB7

+++ php/RemoveUser.php
@@ -7,5 +7,12 @@ if (!$mysqli) {
 }
 
 $email = mysqli_real_escape_string($mysqli, $_GET['email']);
-if ($email != "admin@ehu.es")
+
+if (!isset($_SESSION)) {
+    session_start();
+}
+
+
+
+if (isset($_SESSION['email']) && $_SESSION['email'] == "admin@ehu.es" && $email != "admin@ehu.es")
     $user = mysqli_query($mysqli, "DELETE FROM users WHERE email = '" . $email . "'");
